refactor(memblox): abstract game engine to support multiple spawn modes
This commit refactors the MemBlox engine architecture to support different gameplay mechanics by introducing an abstract base class and specific engine implementations. It also adds a new "Static" game mode where blocks are placed directly rather than falling.

- **Engine Architecture**:
    - Renamed `MemBloxEngine` to `BaseMemBloxEngine` and marked it as `abstract`.
    - Extracted `MemBloxState` and related data models (PowerUps, VFX classes) into a standalone `MemBloxState.kt` file.
    - Introduced `FallingMemBloxEngine`, preserving the original logic where blocks drop from the top of the grid.
    - Introduced `StaticMemBloxEngine`, which calculates the final resting row immediately and implements an "initial reveal" mechanic for new blocks.
- **Logic & State**:
    - Added `initiallyRevealedBlockIds` to `MemBloxState` to track blocks that should be temporarily flipped upon spawning.
    - Updated `MemBloxViewModel` to utilize `StaticMemBloxEngine` for A/B testing.
    - Refined nuke power-up logic to correctly remove specific block IDs from the tracking map.
- **UI Components**:
    - **`MemBloxScreen.kt`**: Updated the block rendering logic to consider `isInitiallyRevealed` when determining if a block should be flipped. Added a 3D flip animation trigger for these initial reveals.
